security: fix CodeQL alerts — integer overflow guards, workflow permissions
- ci.yml: add top-level permissions (contents: read) for all jobs
- rpc: guard uint64→int64 conversions with math.MaxInt64 overflow checks
- rpc: fix uint64→int bounds check using uint64 comparison instead of int cast
- Dismiss SHA-256 crypto alerts as false positives (correct algorithm for
  Lamport OTS and Pedersen commitments)
- Dismiss pion/dtls/v2 dependabot alert (transitive dep from go-ethereum,
  no v2 patch available)

+++ pkg/rpc/api_proof.go
@@ -262,7 +262,7 @@ func (api *EthAPI) getTransactionByBlockHashAndIndex(req *Request) *Response {
 	}
 
 	txs := block.Transactions()
-	if int(index) >= len(txs) {
+	if index >= uint64(len(txs)) {
 		return successResponse(req.ID, nil)
 	}
 
@@ -294,7 +294,7 @@ func (api *EthAPI) getTransactionByBlockNumberAndIndex(req *Request) *Response {
 	}
 
 	txs := block.Transactions()
-	if int(index) >= len(txs) {
+	if index >= uint64(len(txs)) {
 		return successResponse(req.ID, nil)
 	}
